refactor session state store
