+++ frontend/src/lib/components/__tests__/PipelinesPage.test.ts
@@ -1,12 +1,15 @@
 import { render, waitFor } from '@testing-library/svelte';
 import { expect, test, vi } from 'vitest';
 
-vi.mock('$app/stores', () => ({
-  page: {
-    subscribe: (run: any) => { run({ data: { session: { loggedIn: true, org: 'org1' } } }); return () => {}; }
-  }
-}));
+import { sessionStore } from '../../stores/session';
 
+beforeEach(() => {
+  sessionStore.setSession({ loggedIn: true, org: 'org1', userId: null, role: null, csrfToken: null });
+});
+
+afterEach(() => {
+  sessionStore.clear();
+});
 import PipelinesPage from '../../../routes/pipelines/+page.svelte';
 
 const pipelines = [{ id: 'p1', name: 'Pipe', org_id: 'org1', stages: [] }];
